Remove reference to old media file in new media file
