+++ packages/teams-js/src/public/media.ts
@@ -159,121 +159,8 @@ export namespace media {
    * @hidden
    * Hide from docs
    * --------
-   * Base class which holds the callback and notifies events to the host client
-   */
-  export abstract class MediaController<T> {
-    protected controllerCallback: T;
-
-    public constructor(controllerCallback: T) {
-      this.controllerCallback = controllerCallback;
-    }
-
-    protected abstract getMediaType(): MediaType;
-
-    /**
-     * @hidden
-     * Hide from docs
-     * --------
-     * This function will be implemented by the respective media class which holds the logic
-     * of specific events that needs to be notified to the app.
-     * @param mediaEvent indicates the event signed by the host client to the app
-     */
-    protected abstract notifyEventToApp(mediaEvent: MediaControllerEvent): void;
-
-    /**
-     * @hidden
-     * Hide from docs
-     * --------
-     *
-     * Function to notify the host client to programatically control the experience
-     * @param mediaEvent indicates what the event that needs to be signaled to the host client
-     * @returns A promise resolved promise
-     */
-    protected notifyEventToHost(mediaEvent: MediaControllerEvent): Promise<void>;
-    /**
-     * @hidden
-     * Hide from docs
-     * --------
-     *
-     * @deprecated
-     * As of 2.0.0-beta.3, please use {@link media.MediaController.notifyEventToHost media.MediaController.notifyEventToHost(mediaEvent: MediaControllerEvent): Promise\<void\>} instead.
-     *
-     * Function to notify the host client to programatically control the experience
-     * @param mediaEvent indicates what the event that needs to be signaled to the host client
-     * Optional; @param callback is used to send app if host client has successfully handled the notification event or not
-     */
-    protected notifyEventToHost(mediaEvent: MediaControllerEvent, callback?: (err?: SdkError) => void): void;
-    protected notifyEventToHost(mediaEvent: MediaControllerEvent, callback?: (err?: SdkError) => void): Promise<void> {
-      ensureInitialized(FrameContexts.content, FrameContexts.task);
-
-      try {
-        throwExceptionIfMobileApiIsNotSupported(nonFullScreenVideoModeAPISupportVersion);
-      } catch (err) {
-        const wrappedRejectedErrorFn: InputFunction<void> = () => Promise.reject(err);
-
-        return callCallbackWithSdkErrorFromPromiseAndReturnPromise(wrappedRejectedErrorFn, callback);
-      }
-
-      const params: MediaControllerParam = {
-        mediaType: this.getMediaType(),
-        mediaControllerEvent: mediaEvent,
-      };
-
-      const wrappedFunction = (): Promise<void> =>
-        new Promise(resolve => resolve(sendAndHandleSdkError('media.controller', [params])));
-
-      return callCallbackWithSdkErrorFromPromiseAndReturnPromise(wrappedFunction, callback);
-    }
-
-    /**
-     * Function to programatically stop the ongoing media event
-     *
-     * @returns A resolved promise
-     * */
-    public stop(): Promise<void>;
-    /**
-     *
-     * Function to programatically stop the ongoing media event
-     *
-     * @deprecated
-     * As of 2.0.0-beta.3, please use {@link media.MediaController.stop media.MediaController.stop(): Promise\<void\>} instead.
-     *
-     * Optional; @param callback is used to send app if host client has successfully stopped the event or not
-     */
-    public stop(callback?: (err?: SdkError) => void): void;
-    public stop(callback?: (err?: SdkError) => void): Promise<void> {
-      return Promise.resolve(this.notifyEventToHost(MediaControllerEvent.StopRecording, callback));
-    }
-  }
-
-  /**
-   * @hidden
-   * Hide from docs
-   * --------
    * Events which are used to communicate between the app and the host client during the media recording flow
    */
-  export enum MediaControllerEvent {
-    StartRecording = 1,
-    StopRecording = 2,
-  }
-
-  /**
-   * @hidden
-   * Hide from docs
-   * --------
-   * Interface with relevant info to send communication from the app to the host client
-   */
-  interface MediaControllerParam {
-    /**
-     * List of team information
-     */
-    mediaType: media.MediaType;
-
-    /**
-     * List of team information
-     */
-    mediaControllerEvent: MediaControllerEvent;
-  }
 
   /**
    * @hidden
